i18n: Wrap title input on a new line when title examples can't fit
