+++ client/landing/gutenboarding/onboarding-block/acquire-intent/acquire-intent-text-input/index.tsx
@@ -1,5 +1,6 @@
 import classnames from 'classnames';
 import * as React from 'react';
+import getTextWidth from '../get-text-width';
 
 import './style.scss';
 
@@ -14,32 +15,6 @@ interface Props {
 	placeholder?: string;
 	id?: string;
 	forwardedRef: React.MutableRefObject< HTMLInputElement | undefined >;
-}
-
-/**
- * A canvas to use to create text and measure its width.
- * This is needed for the underline width.
- */
-const textSizingCanvas = document.createElement( 'canvas' );
-textSizingCanvas.width = 2000;
-textSizingCanvas.height = 100;
-const canvasContext = textSizingCanvas.getContext( '2d' ) as CanvasRenderingContext2D;
-
-/**
- * Gets the font info from an input field then measures the width of the text within
- *
- * @param text the string
- * @param element The input element
- */
-function getTextWidth( text: string, element: HTMLInputElement | undefined ) {
-	if ( ! element || ! text ) {
-		return 0;
-	}
-	const computedCSS = window.getComputedStyle( element );
-
-	// FF returns an empty strong in font prop
-	canvasContext.font = computedCSS.font || `${ computedCSS.fontSize } ${ computedCSS.fontFamily }`;
-	return canvasContext.measureText( text ).width;
 }
 
 const AcquireIntentTextInput: React.FunctionComponent< Props > = ( {
