+++ client/landing/gutenboarding/onboarding-block/acquire-intent/site-title.tsx
@@ -3,11 +3,13 @@ import { Icon } from '@wordpress/icons';
 import { useI18n } from '@wordpress/react-i18n';
 import classnames from 'classnames';
 import * as React from 'react';
+import { useWindowResizeCallback } from 'calypso/lib/track-element-size';
 import useTyper from '../../hooks/use-typer';
 import { recordSiteTitleSelection } from '../../lib/analytics';
 import { useIsAnchorFm } from '../../path';
 import { STORE_KEY } from '../../stores/onboard';
 import AcquireIntentTextInput from './acquire-intent-text-input';
+import getTextWidth from './get-text-width';
 import tip from './tip';
 
 interface Props {
@@ -21,47 +23,66 @@ const SiteTitle: React.FunctionComponent< Props > = ( { onSubmit, inputRef } ) =
 	const isAnchorFmSignup = useIsAnchorFm();
 	const { setSiteTitle } = useDispatch( STORE_KEY );
 	const [ isTouched, setIsTouched ] = React.useState( false );
-	const siteTitleExamples = [
-		/* translators: This is an example of a site name,
+	const siteTitleExamples = React.useMemo(
+		() => [
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'The Local Latest', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'The Local Latest', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'North Peak Cycling', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'North Peak Cycling', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Sunshine Daycare', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Sunshine Daycare', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Quick Wins Consulting', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Quick Wins Consulting', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Puns and Pedantry', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Puns and Pedantry', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Yoga For Everyone', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Yoga For Everyone', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Pugs Wearing Bowties', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Pugs Wearing Bowties', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Behind the Lens', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Behind the Lens', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Marketing Magic', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Marketing Magic', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Cortado Coffee', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Cortado Coffee', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Mumbai Bites', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'Mumbai Bites', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'RPM Motors', 'sample site title' ),
-		/* translators: This is an example of a site name,
+			_x( 'RPM Motors', 'sample site title' ),
+			/* translators: This is an example of a site name,
 		   feel free to create your own but please keep it under 22 characters */
-		_x( 'Max’s Burger Bar', 'sample site title' ),
-	];
+			_x( 'Max’s Burger Bar', 'sample site title' ),
+		],
+		[ _x ]
+	);
+	const inputElement = inputRef.current;
+	const maxTitleWidth = React.useMemo(
+		() =>
+			siteTitleExamples.reduce(
+				( maxWidth, title ) => Math.max( maxWidth, getTextWidth( title, inputElement ) ),
+				0
+			),
+		[ siteTitleExamples, inputElement ]
+	);
+	const [ hasOverflowingPlaceholder, setHasOverflowingPlaceholder ] = React.useState( false );
+	const labelRef = React.useRef();
+	const resizeRef = useWindowResizeCallback( ( formDomRect ) => {
+		const labelDomRect = labelRef.current.getBoundingClientRect();
+
+		setHasOverflowingPlaceholder( maxTitleWidth > formDomRect.width - labelDomRect.width );
+	} );
 
 	const handleFormSubmit = ( e: React.FormEvent< HTMLFormElement > ) => {
 		// hitting 'Enter' when focused on the input field should direct to next step.
@@ -83,10 +104,12 @@ const SiteTitle: React.FunctionComponent< Props > = ( { onSubmit, inputRef } ) =
 
 	return (
 		<form
+			ref={ resizeRef }
 			className={ classnames( 'site-title', { 'is-touched': isTouched } ) }
 			onSubmit={ handleFormSubmit }
 		>
 			<label
+				ref={ labelRef }
 				htmlFor="site-title__input"
 				className="site-title__input-label"
 				data-e2e-string="My site is called"
@@ -94,7 +117,11 @@ const SiteTitle: React.FunctionComponent< Props > = ( { onSubmit, inputRef } ) =
 				{ /* translators: label for site title input in Gutenboarding */ }
 				{ isAnchorFmSignup ? __( 'My podcast is called' ) : __( 'My site is called' ) }
 			</label>
-			<div className="site-title__input-wrapper">
+			<div
+				className={ classnames( 'site-title__input-wrapper', {
+					'has-overflowing-placeholder': hasOverflowingPlaceholder,
+				} ) }
+			>
 				{ /* Adding key makes it more performant
 					because without it the element is recreated
 					for every letter in the typing animation
